refactor: move layout of image commands to match v1
